remove ingestion tests
