+++ integration_tests/utils.py
@@ -10,7 +10,6 @@ from datetime import datetime
 from pathlib import Path
 import numpy as np
 import rasterio
-import yaml
 from click.testing import CliRunner
 
 from datacube.utils.documents import load_from_yaml
@@ -75,55 +74,6 @@ def limit_num_measurements(dataset_type):
     if len(measurements) > TEST_STORAGE_NUM_MEASUREMENTS:
         dataset_type['measurements'] = measurements[:TEST_STORAGE_NUM_MEASUREMENTS]
     return dataset_type
-
-
-def prepare_test_ingestion_configuration(tmpdir,
-                                         output_dir,
-                                         filename,
-                                         mode=None):
-    customizers = {
-        'fast_ingest': edit_for_fast_ingest,
-        'end2end': edit_for_end2end,
-    }
-
-    filename = Path(filename)
-    if output_dir is None:
-        output_dir = tmpdir.ensure(filename.stem, dir=True)
-    config = load_yaml_file(filename)[0]
-
-    if mode is not None:
-        if mode not in customizers:
-            raise ValueError('Wrong mode: ' + mode)
-        config = customizers[mode](config)
-
-    config['location'] = str(output_dir)
-
-    # If ingesting with the s3test driver
-    if 'bucket' in config['storage']:
-        config['storage']['bucket'] = str(output_dir)
-
-    config_path = tmpdir.join(filename.name)
-    with open(str(config_path), 'w') as stream:
-        yaml.safe_dump(config, stream)
-    return config_path, config
-
-
-def edit_for_end2end(config):
-    storage = config.get('storage', {})
-
-    storage['crs'] = 'EPSG:3577'
-    storage['tile_size']['x'] = 100000.0
-    storage['tile_size']['y'] = 100000.0
-
-    config['storage'] = storage
-    return config
-
-
-def edit_for_fast_ingest(config):
-    config = alter_product_for_testing(config)
-    config['storage']['crs'] = 'EPSG:28355'
-    config['storage']['chunking']['time'] = 1
-    return config
 
 
 def _make_geotiffs(tiffs_dir, day_offset, num_bands=NUM_BANDS):
@@ -243,3 +193,10 @@ def alter_product_for_testing(product, metadata_type=None):
     if metadata_type:
         product['metadata_type'] = metadata_type.name
     return product
+
+
+def ensure_datasets_are_indexed(index, valid_uuids):
+    datasets = index.datasets.search_eager(product='ls5_nbar_scene')
+    assert len(datasets) == len(valid_uuids)
+    for dataset in datasets:
+        assert dataset.id in valid_uuids
